Move numba/dppl_config.py to numba_dppy/config.py
dppl_config.py contains flag `dppl_present` which indicate that `dpCtl`
and devices are available.
As `numba-dppy` will depend on `dpCtl` then the check is not necessary.
Also checking for available platform and device could be done in
`numba-dppy`. It also can raise and exception on import as before.

`dppl_present` also renamed to `dppy_present`.