Add setup.py and conda-recipe (#5)
* Add setup.py

It is initial draft version.

* Replace numba.dppl with numba_dppy

Now numba.dppl is placed in separate package numba_dppy.

* Import numba_dppy as dppl

Previousply it was `from numba import dppl`.
Now it is `import numba_dppy as dppl`.

* Import numba_dppy with its name and as dppl

There are many places in code which access to `numba_dppy`.
Previously it was `numba.dppl`. It was accessible because
`from numba import dppl`.

* Add version, url, author and classifiers to setup.py

This parameters are necessary for PyPI.

* Add conda-recipe

* Use dppy in dispatcher_registry and TargetDispatcher

`dispatcher_registry.ondemand['dppl'] = init_jit` can be used only one
time. So `dppl` was renamed to `dppy`.
Also in class TargetDispatcher field `target_dppl = 'dppl'` renamed to
`dppy`.

* Use --single-version-externally-managed

This flag configure installation of `numba_dppy` as folder in
`site-packages`. The folder is required for installing compiled spirv
code to it.

* Add compilation of SPIRV code

* Add dependency to numba and dpctl in conda recipe

* Add MANIFEST.in